new_project/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.6.0
dj-database-url==2.2.0
Pillow==10.4.0
